frontend/Interface/interactive-quiz/requirements.txt
Flask==2.1.0
Flask-CORS==3.0.10
mysql-connector-python==8.0.27